refactor: minor review comments (cleaning clutter to get to the juicy stuff)
