+++ nodebuilder/header/module.go
@@ -6,12 +6,11 @@ import (
 	logging "github.com/ipfs/go-log/v2"
 	"go.uber.org/fx"
 
+	"github.com/celestiaorg/celestia-node/fraud"
+	"github.com/celestiaorg/celestia-node/header"
 	"github.com/celestiaorg/celestia-node/header/p2p"
 	"github.com/celestiaorg/celestia-node/header/store"
 	"github.com/celestiaorg/celestia-node/header/sync"
-
-	"github.com/celestiaorg/celestia-node/fraud"
-	"github.com/celestiaorg/celestia-node/header"
 	"github.com/celestiaorg/celestia-node/libs/fxutil"
 	"github.com/celestiaorg/celestia-node/nodebuilder/node"
 )
@@ -84,6 +83,6 @@ func Module(tp node.Type, cfg *Config) fx.Option {
 			fx.Supply(header.MakeExtendedHeader),
 		)
 	default:
-		panic("wrong node type")
+		panic("invalid node type")
 	}
 }
